Move filter queries utils to own file
